stress-dup: remove redundant double cast on a double type

+++ stress-dup.c
@@ -427,7 +427,7 @@ tidy_mmap:
 		(void)munmap((void *)info, sizeof(*info));
 	}
 #endif
-	rate = (dup_count > 0.0) ? (double)dup_duration / dup_count : 0.0;
+	rate = (dup_count > 0.0) ? dup_duration / dup_count : 0.0;
 	stress_metrics_set(args, "nanosecs per dup call",
 		rate * STRESS_DBL_NANOSECOND, STRESS_METRIC_HARMONIC_MEAN);
 	stress_metrics_set(args, "dup calls",
